pyproject.toml
[tool.poetry]
name = "bestvideo"
version = "0.1.0"
description = "BestVideo: Video processing and selection tool"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
typer = {extras = ["all"], version = "^0.9.0"} # Updated to a more recent version often used with Pydantic v2
pydantic = "^2.5.3" # Compatible with Typer and common usage
rich = "^13.7.0"
tqdm = "^4.66.1"
PyYAML = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[tool.poetry.scripts]
bestvideo = "cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 